+++ src/Core/Application/Catalog/Products/UpdateProductRequestValidator.cs
@@ -5,11 +5,10 @@ public class UpdateProductRequestValidator : CustomValidator<UpdateProductReques
     public UpdateProductRequestValidator(IReadRepository<Product> productRepo, IReadRepository<Brand> brandRepo, IStringLocalizer<UpdateProductRequestValidator> T)
     {
         RuleFor(p => p.Name)
-            .NotEmpty()
             .MaximumLength(75)
             .MustAsync(async (product, name, ct) =>
                     await productRepo.FirstOrDefaultAsync(new ProductByNameSpec(name), ct)
-                        is not Product existingProduct || existingProduct.Id == product.Id)
+                        is not Product existingProduct || existingProduct?.Id == product.Id)
                 .WithMessage((_, name) => T["Product {0} already Exists.", name]);
 
         RuleFor(p => p.Price)