+++ src/Host/Controllers/Identity/UsersController.cs
@@ -1,0 +1,72 @@
+using Tekton.API.Application.Identity.Users;
+using Tekton.API.Application.Identity.Users.Password;
+
+namespace Tekton.API.Host.Controllers.Identity;
+
+public class UsersController : VersionNeutralApiController
+{
+    private readonly IUserService _userService;
+
+    public UsersController(IUserService userService) => _userService = userService;
+
+    [HttpGet]
+    [MustHavePermission(TektonAction.View, TektonResource.Users)]
+    [OpenApiOperation("Get list of all users.", "")]
+    public Task<List<UserDetailsDto>> GetListAsync(CancellationToken cancellationToken)
+    {
+        return _userService.GetListAsync(cancellationToken);
+    }
+
+    [HttpGet("{id}")]
+    [MustHavePermission(TektonAction.View, TektonResource.Users)]
+    [OpenApiOperation("Get a user's details.", "")]
+    public Task<UserDetailsDto> GetByIdAsync(string id, CancellationToken cancellationToken)
+    {
+        return _userService.GetAsync(id, cancellationToken);
+    }
+
+    [HttpGet("{id}/roles")]
+    [MustHavePermission(TektonAction.View, TektonResource.UserRoles)]
+    [OpenApiOperation("Get a user's roles.", "")]
+    public Task<List<UserRoleDto>> GetRolesAsync(string id, CancellationToken cancellationToken)
+    {
+        return _userService.GetRolesAsync(id, cancellationToken);
+    }
+
+    [HttpPost("{id}/roles")]
+    [ApiConventionMethod(typeof(TektonApiConventions), nameof(TektonApiConventions.Register))]
+    [MustHavePermission(TektonAction.Update, TektonResource.UserRoles)]
+    [OpenApiOperation("Update a user's assigned roles.", "")]
+    public Task<string> AssignRolesAsync(string id, UserRolesRequest request, CancellationToken cancellationToken)
+    {
+        return _userService.AssignRolesAsync(id, request, cancellationToken);
+    }
+
+    [HttpPost]
+    [MustHavePermission(TektonAction.Create, TektonResource.Users)]
+    [OpenApiOperation("Creates a new user.", "")]
+    public Task<string> CreateAsync(CreateUserRequest request)
+    {
+        // TODO: check if registering anonymous users is actually allowed (should probably be an appsetting)
+        // and return UnAuthorized when it isn't
+        // Also: add other protection to prevent automatic posting (captcha?)
+        return _userService.CreateAsync(request, GetOriginFromRequest());
+    }
+
+    [HttpPost("{id}/toggle-status")]
+    [MustHavePermission(TektonAction.Update, TektonResource.Users)]
+    [ApiConventionMethod(typeof(TektonApiConventions), nameof(TektonApiConventions.Register))]
+    [OpenApiOperation("Toggle a user's active status.", "")]
+    public async Task<ActionResult> ToggleStatusAsync(string id, ToggleUserStatusRequest request, CancellationToken cancellationToken)
+    {
+        if (id != request.UserId)
+        {
+            return BadRequest();
+        }
+
+        await _userService.ToggleStatusAsync(request, cancellationToken);
+        return Ok();
+    }
+
+    private string GetOriginFromRequest() => $"{Request.Scheme}://{Request.Host.Value}{Request.PathBase.Value}";
+}