+++ src/Infrastructure/Persistence/Initialization/DatabaseInitializer.cs
@@ -32,14 +32,14 @@ internal class DatabaseInitializer : IDatabaseInitializer
         _logger.LogInformation("For documentations and guides, write to José Angel Bustamante");
     }
 
-    public async Task InitializeApplicationDbForTenantAsync(FSHTenantInfo tenant, CancellationToken cancellationToken)
+    public async Task InitializeApplicationDbForTenantAsync(TektonTenantInfo tenant, CancellationToken cancellationToken)
     {
         // First create a new scope
         using var scope = _serviceProvider.CreateScope();
 
         // Then set current tenant so the right connectionstring is used
         _serviceProvider.GetRequiredService<IMultiTenantContextAccessor>()
-            .MultiTenantContext = new MultiTenantContext<FSHTenantInfo>()
+            .MultiTenantContext = new MultiTenantContext<TektonTenantInfo>()
             {
                 TenantInfo = tenant
             };
@@ -64,7 +64,7 @@ internal class DatabaseInitializer : IDatabaseInitializer
     {
         if (await _tenantDbContext.TenantInfo.FindAsync(new object?[] { MultitenancyConstants.Root.Id }, cancellationToken: cancellationToken) is null)
         {
-            var rootTenant = new FSHTenantInfo(
+            var rootTenant = new TektonTenantInfo(
                 MultitenancyConstants.Root.Id,
                 MultitenancyConstants.Root.Name,
                 string.Empty,