+++ src/Infrastructure/Persistence/Initialization/IDatabaseInitializer.cs
@@ -5,5 +5,5 @@ namespace Tekton.API.Infrastructure.Persistence.Initialization;
 internal interface IDatabaseInitializer
 {
     Task InitializeDatabasesAsync(CancellationToken cancellationToken);
-    Task InitializeApplicationDbForTenantAsync(FSHTenantInfo tenant, CancellationToken cancellationToken);
+    Task InitializeApplicationDbForTenantAsync(TektonTenantInfo tenant, CancellationToken cancellationToken);
 }