+++ src/Migrators/Migrators.MSSQL/Migrations/Tenant/20220125214527_InitialMigrations.Designer.cs
@@ -24,7 +24,7 @@ namespace Migrators.MSSQL.Migrations.Tenant
 
             SqlServerModelBuilderExtensions.UseIdentityColumns(modelBuilder, 1L, 1);
 
-            modelBuilder.Entity("Tekton.API.Infrastructure.Multitenancy.FSHTenantInfo", b =>
+            modelBuilder.Entity("Tekton.API.Infrastructure.Multitenancy.TektonTenantInfo", b =>
                 {
                     b.Property<string>("Id")
                         .HasMaxLength(64)