+++ src/Migrators/Migrators.Oracle/Migrations/Tenant/TenantDbContextModelSnapshot.cs
@@ -22,7 +22,7 @@ namespace Migrators.Oracle.Migrations.Tenant
 
             OracleModelBuilderExtensions.UseIdentityColumns(modelBuilder, 1L, 1);
 
-            modelBuilder.Entity("Tekton.API.Infrastructure.Multitenancy.FSHTenantInfo", b =>
+            modelBuilder.Entity("Tekton.API.Infrastructure.Multitenancy.TektonTenantInfo", b =>
                 {
                     b.Property<string>("Id")
                         .HasMaxLength(64)